% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catches.R
\docType{data}
\name{catches}
\alias{catches}
\title{Sample data for fisheries_summary function}
\format{A matrix with rownames as fish species names and columnnames as fishery names}
\usage{
catches
}
\description{
catches - A matrix with number of each fish species caught by fishery
}
\keyword{datasets}
