% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisheries_summary.R
\name{fisheries_summary}
\alias{fisheries_summary}
\title{Summary of Fisheries Data}
\usage{
fisheries_summary(catches, prices, plot = FALSE)
}
\arguments{
\item{catches}{a matrix containing the number of each fish species caught in each location,
with rows for species and columns for locations}

\item{prices}{a data frame containing the price of each fish species}

\item{plot}{if set to TRUE, will generate a plot of revenue by fishery}
}
\description{
This function computes the most frequently caught fish in each fishery,
the revenue from each fishery, the total revenue from all fisheries,
and optionally generates a plot of revenue by fishery.
}
\author{
Kenneth Larrieu
}
