% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_rate.R
\name{growth_rate}
\alias{growth_rate}
\title{Fish Growth Rate}
\usage{
growth_rate(t, a = -0.497, b = 0.1656, c = 0.08588, d = -0.004266)
}
\arguments{
\item{t}{temperature in degrees Celsius}

\item{a}{constant term in estimate function}

\item{b}{first degree coefficient in estimate function}

\item{c}{second degree coefficient in estimate function}

\item{d}{third degree coefficient in estimate function}
}
\description{
This function estimates the growth rate per day for Atlantic Cod, using a third
degree polynomial function of temperature (in degrees Celsius).
The default values of parameters a,b,c,d are obtained from
Experiment A in Table 2 of Bjoornsson et al.
}
\examples{
growth_rate(15)
}
\references{
Bjoornsson et al. Growth model for Atlantic cod (Gadus morhua):
Effects of temperature and body weight on growth rate.
}
\author{
Kenneth Larrieu
}
