% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices.R
\docType{data}
\name{prices}
\alias{prices}
\title{Sample data for fisheries_summary function}
\format{A data frame with two columns:
\itemize{
\item fish: name of fish species
\item price: going price of the species}}
\usage{
prices
}
\description{
prices - A data frame with prices of each fish species
}
\keyword{datasets}
